Make web tests work better
